embed_service/requirements.txt
grpcio
grpcio-tools
FlagEmbedding
fastapi
pydantic
uvicorn